chore: move validation of FC metrics to host_tools
test_cli_metrics_path and test_flush_metrics both validated breaking
change in FC metrics so move the validation functions to host_tools
which already had CI metrics related functions defined.
